+++ example-project/maven/README.md
@@ -1,0 +1,1 @@
+Maven repo directory
