+++ gradle-embed-maven-repo/build.gradle
@@ -1,4 +1,5 @@
 apply plugin: "groovy"
+apply plugin: 'maven'
 
 buildscript {
 	repositories {
@@ -22,7 +23,7 @@ dependencies {
 	compile localGroovy() 
 }
 
-version = "1.0.0"
+version = "1.0.6"
 group = "com.mageddo"
 
 pluginBundle {
