+++ gradle-embed-maven-repo/src/main/resources/META-INF/gradle-plugins/com.mageddo.gradle-embed-maven-repo.properties
@@ -1,0 +1,1 @@
+implementation-class = com.mageddo.repo.EmbedMavenRepoPlugin
